+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - TDD
 - db/functions の引数
 - reagent, マップの場所はそこか？
+- goods received 表示だけ abbrev
+
 
 ## 0.9.0 - 2022-06-02
 - l.melt にデプロイ